Added entity etc

+++ src/main/java/SirQuizALot/SirQuizALotController.java
@@ -132,7 +132,7 @@ public class SirQuizALotController {
     }
 
     @PostMapping("/newquestion")
-    public String addQuestion(@RequestParam int id, @RequestParam String question, @RequestParam String alternative1, @RequestParam String alternative2, @RequestParam String alternative3, @RequestParam int answer, HttpSession session) {
+    public String addQuestion(@RequestParam Long id, @RequestParam String question, @RequestParam String alternative1, @RequestParam String alternative2, @RequestParam String alternative3, @RequestParam int answer, HttpSession session) {
         String username = (String) session.getAttribute("username");
         service.createQuestion(id, question, alternative1, alternative2, alternative3, answer);
         return "redirect:/admin";
@@ -144,7 +144,7 @@ public class SirQuizALotController {
     }
 
     @PostMapping("/requestnewquestion")
-    public String requestQuestion2(@RequestParam int id, @RequestParam String question, @RequestParam String alt1, @RequestParam String alt2, @RequestParam String alt3, @RequestParam int answer) {
+    public String requestQuestion2(@RequestParam Long id, @RequestParam String question, @RequestParam String alt1, @RequestParam String alt2, @RequestParam String alt3, @RequestParam int answer) {
         service.questionRequest(id, question, alt1, alt2, alt3, answer);
         return "redirect:/home";
